Remove InferBIDSURIs.
